MSP-22 Moved DTOs to another folder
