+++ src/main/java/de/gaknr/mspbackend/clothingitem/ClothingItemController.java
@@ -1,11 +1,11 @@
 package de.gaknr.mspbackend.clothingitem;
 
+import de.gaknr.mspbackend.clothingitem.dtos.GetClothingItemDTO;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.bson.types.ObjectId;
 import org.springframework.http.HttpStatus;
