+++ src/main/java/de/gaknr/mspbackend/clothingitem/ClothingItemMapper.java
@@ -1,5 +1,7 @@
 package de.gaknr.mspbackend.clothingitem;
 
+import de.gaknr.mspbackend.clothingitem.dtos.AddClothingItemDTO;
+import de.gaknr.mspbackend.clothingitem.dtos.GetClothingItemDTO;
 import org.springframework.stereotype.Service;
 
 import java.util.Base64;
