+++ src/main/java/de/gaknr/mspbackend/clothingitem/dtos/GetClothingItemDTO.java
@@ -1,4 +1,4 @@
-package de.gaknr.mspbackend.clothingitem;
+package de.gaknr.mspbackend.clothingitem.dtos;
 
 import de.gaknr.mspbackend.clothingitem.enums.*;
 import org.bson.types.ObjectId;
